+++ README.md
@@ -49,7 +49,8 @@ python app/game.py
 Run the inventory report:
 
 ```sh
-python app/groceries.py
+#python app/groceries.py
+python -m app.groceries
 ```
 
 
@@ -59,11 +60,13 @@ python app/groceries.py
 Run the stocks report:
 
 ```sh
-python app/stocks.py
+#python app/stocks.py
+python -m app.stocks
 ```
 
 Run the crypto report:
 
 ```sh
-python app/crypto.py
+#python app/crypto.py
+python -m app.crypto
 ```
