+++ app/groceries.py
@@ -7,26 +7,7 @@
 
 import os
 
-
-def to_usd(my_price):
-    """
-    This is a docstring. It tells us what this function is about.
-    What its responsibilities ar.
-    What the params are ab9out.
-    What datatypes the params are.
-    What this function will return.
-    Example of invoking the function.
-
-    Invoke like this: to_usd(9.9999)
-    """
-    return '${:,.2f}'.format(my_price)
-
-
-
-
-
-
-
+from app.utils import to_usd
 
 
 # checks to see if a products.csv file exists. If not, it uses the default
