+++ app/stocks.py
@@ -6,6 +6,8 @@ import os
 from dotenv import load_dotenv
 from pandas import read_csv
 
+from app.utils import to_usd
+
 load_dotenv()
 
 ALPHAVANTAGE_API_KEY = os.getenv("ALPHAVANTAGE_API_KEY", default="demo")
@@ -22,4 +24,4 @@ latest = df.iloc[0]
 print(symbol)
 print(latest["timestamp"])
 print(latest["close"])
-print('${:,.2f}'.format(latest["close"]))
+print(to_usd(latest["close"]))
